requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.7.0
aiohttp
aiormq==3.3.1
apscheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.4
boto3==1.19.7
botocore==1.22.7
bottle==0.12.19
cachecontrol==0.12.10
cachetools==4.2.4
cffi==1.15.0
chardet==3.0.4
clang==5.0
click==8.0.3
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.0
coloredlogs==15.0.1
colorhash==1.0.3
crypto==1.4.1
cryptography==3.4.8
cycler==0.11.0
decorator==4.4.2
deprecated==1.2.13
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
dominate==2.6.0
fbmessenger==6.0.0
flask==0.12.5
flask-appconfig==0.11.1
flask-bootstrap==3.3.7.1
flask-cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gcloud==0.18.3
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
httpcore
httplib2==0.20.1
httptools==0.3.0
httpx
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
jinja2==3.0.3
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
jwcrypto==1.0
kafka-python==2.0.2
keras==2.6.0
keras-preprocessing==1.1.2
kiwisolver==1.3.2
markdown==3.3.4
markupsafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.3
multidict
naked==0.1.31
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pillow==8.4.0
prompt-toolkit==2.0.10
proto-plus==1.20.1
protobuf==3.19.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pyjwt==2.0.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==2.4.7
pyreadline3==3.3
pyrsistent==0.18.0
pytelegrambotapi==3.8.3
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-engineio==4.3.0
python-firebase
python-jwt==3.3.2
python-socketio==5.4.1
pytz==2021.3
pyyaml==6.0
questionary==1.9.0
rasa==2.6.2
rasa-sdk==2.8.4
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-async
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-api==1.15.0
rsa==4.7.2
ruamel-yaml==0.16.13
ruamel-yaml-clib==0.2.6
s3transfer==0.5.0
sanic==20.9.1
sanic-cors==0.10.0.post3
sanic-jwt==1.6.0
sanic-plugins-framework==0.9.5
scikit-learn==0.24.2
scipy==1.7.1
sentry-sdk==0.19.5
shellescape==3.8.1
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
sniffio==1.2.0
sqlalchemy==1.3.24
sseclient==0.0.27
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.4
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==3.0.0
tqdm==4.59.0
twilio==6.50.1
typeguard==2.13.0
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.2.0
uritemplate==4.1.1
urllib3==1.26.7
visitor==0.1.3
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
werkzeug==0.16.1
wrapt==1.12.1
yarl==1.7.0
